feat: add validation patterns
