+++ react-components/src/components/pages/FormPage/functions/validateDate.ts
@@ -1,0 +1,16 @@
+export function validateDate(date: string): boolean {
+  const birthDateMs = Date.parse(date);
+  const currentDate = new Date().toISOString();
+  const currentDateMs = Date.parse(currentDate);
+  const years18 = 18 * 1000 * 60 * 60 * 24 * 365;
+  const years100 = 100 * 1000 * 60 * 60 * 24 * 365;
+  let noNewerThan18Years = true;
+  if (currentDateMs - birthDateMs < years18) {
+    noNewerThan18Years = false;
+  }
+  let noElderThan100Years = true;
+  if (currentDateMs - birthDateMs > years100) {
+    noElderThan100Years = false;
+  }
+  return noNewerThan18Years && noElderThan100Years;
+}
